% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema.R
\name{fema_disaster}
\alias{fema_disaster}
\title{Get FEMA disaster summary by disaster number}
\usage{
fema_disaster(disaster_number)
}
\arguments{
\item{disaster_number}{Integer disaster number}
}
\value{
tibble: single row with disaster details
}
\description{
Get FEMA disaster summary by disaster number
}
