% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema.R
\name{fema_disasters}
\alias{fema_disasters}
\title{Search FEMA disaster declarations}
\usage{
fema_disasters(
  top = 100,
  orderby = "declarationDate desc",
  state = NULL,
  incident_type = NULL
)
}
\arguments{
\item{top}{Maximum number of records to return (default 100)}

\item{orderby}{OData orderby clause (default "declarationDate desc")}

\item{state}{Optional two-letter state abbreviation filter}

\item{incident_type}{Optional incident type filter (e.g. "Fire", "Hurricane")}
}
\value{
tibble: disaster_number, state, declaration_type, declaration_date,
incident_type, title, incident_begin, incident_end, designated_area
}
\description{
Search FEMA disaster declarations
}
