% Generated by build script
\name{fema_disaster}
\alias{fema_disaster}
\title{Get FEMA disaster summary by disaster number}
\usage{
fema_disaster()
}
\arguments{
\item{disaster_number}{Integer disaster number}
}
\value{tibble: single row with disaster details}
\description{Get FEMA disaster summary by disaster number}
