% Generated by build script
\name{fema_disasters}
\alias{fema_disasters}
\title{Search FEMA disaster declarations}
\usage{
fema_disasters()
}
\arguments{
\item{top}{Maximum number of records to return (default 100)}
\item{orderby}{OData orderby clause (default "declarationDate desc")}
\item{state}{Optional two-letter state abbreviation filter}
\item{incident_type}{Optional incident type filter (e.g. "Fire", "Hurricane")}
}
\value{tibble: disaster_number, state, declaration_type, declaration_date,}
\description{Search FEMA disaster declarations}
